deal with new `ImTextureID` type
in https://github.com/ocornut/imgui/commit/92b94980c69ff3d3d7f5dc30c1c19abfb72db47e `ImTextureID` was updated from `void*` to `ImU64`, which is defined as `unsigned long long` (which conflicts with some implementations of `uint64_t` that have it as `unsigned long`)

+++ src/graphic/Fast3D/gfx_direct3d11.cpp
@@ -1229,7 +1229,7 @@ gfx_d3d11_get_pixel_depth(int fb_id, const std::set<std::pair<float, float>>& co
 
 } // namespace
 
-ImTextureID gfx_d3d11_get_texture_by_id(int id) {
+ImTextureID gfx_d3d11_get_texture_by_id(unsigned long long id) {
     return d3d.textures[id].resource_view.Get();
 }
