+++ src/window/gui/resource/GuiTexture.h
@@ -7,7 +7,7 @@ namespace Ship {
 #define RESOURCE_TYPE_GUI_TEXTURE 0x47544558 // GTEX
 
 struct GuiTextureMetadata {
-    uint32_t RendererTextureId;
+    unsigned long long RendererTextureId;
     int32_t Width;
     int32_t Height;
 };
